Fix "copy and paste" error in documentation
Split FixturesAwareTestCase tests in several classes, one for each type

Remove code coverage ignore directives from FixturesAwareTestCase and add tests to cover those parts
